Part-2/TaskScheduler/backend/services/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Authentication microservice – handles user registration, login, JWT tokens and API keys"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.1.6",
    "email-validator>=2.0.0",
    "fastapi[standard]>=0.116.0",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.7.1",
    "PyJWT>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.1.0",
    "asgi-lifespan>=2.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
